# openapi.json
{"openapi":"3.1.0","info":{"title":"LegalTech AI Contract Scanner API","description":"AI-powered legal contract analysis platform","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/webhooks/clerk":{"post":{"tags":["auth"],"summary":"Clerk Webhook","description":"Handle Clerk webhooks for user sync.","operationId":"clerk_webhook_api_v1_webhooks_clerk_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contracts/":{"get":{"tags":["contracts"],"summary":"Get Contracts","description":"Return a list of all contracts for the authenticated user.\nIncludes overall risk score from analysis_results.","operationId":"get_contracts_api_v1_contracts__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/contracts/{contractId}":{"get":{"tags":["contracts"],"summary":"Get Contract Detail","description":"Return the full contract detail including all clause results,\nanalysis result, and scan job status.","operationId":"get_contract_detail_api_v1_contracts__contractId__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contractId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Contractid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["contracts"],"summary":"Delete Contract","description":"Hard-delete the contract and all associated data for the authenticated user.","operationId":"delete_contract_api_v1_contracts__contractId__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contractId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Contractid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/upload/":{"post":{"tags":["upload"],"summary":"Upload Contract","description":"Upload a contract and trigger a scan job.\nRate limited to 10 uploads per hour.","operationId":"upload_contract_api_v1_upload__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContractCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/scan/{contractId}":{"post":{"tags":["analysis"],"summary":"Trigger Scan","description":"Manually trigger or retrigger a scan for a contract the user owns.\nIf a scan is already complete, return the existing results.\nIf failed, reset and requeue.","operationId":"trigger_scan_api_v1_scan__contractId__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contractId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Contractid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scan/{jobId}":{"get":{"tags":["analysis"],"summary":"Get Scan Status","description":"Return the current ScanJobStatus (status, progress_pct, error_message).\nVerify JWT and ownership.","operationId":"get_scan_status_api_v1_scan__jobId__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Jobid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scan/{job_id}/stream":{"get":{"tags":["streaming"],"summary":"Stream clause results for a scan job via SSE","description":"Opens a persistent SSE connection for ``job_id``.\n\nSecurity:\n- ``get_current_user`` validates the JWT from the ``Authorization`` header.\n- Ownership of the job is verified inside ``_sse_generator`` against the DB.\n\nThe client receives:\n- ``data: {...}``           — clause result events (default event type)\n- ``event: progress``       — progress_pct updates\n- ``event: complete``       — signals the scan is done; client should close\n- ``event: error``          — unrecoverable pipeline failure\n- ``: heartbeat``           — SSE comment every 15 s (no browser event fired)","operationId":"stream_scan_results_api_v1_scan__job_id__stream_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"text/event-stream","content":{"application/json":{"schema":{}}}},"401":{"description":"Invalid or missing JWT"},"404":{"description":"Job not found or does not belong to user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/counter-offer/{clause_id}":{"post":{"tags":["counter-offer"],"summary":"Generate Counter Offer","description":"Triggers counter-offer generation for a HIGH-risk clause.\n\n- Verifies JWT and ownership\n- Checks if counter-offer already exists (returns it if so)\n- Otherwise queues the generate_counter_offer Celery task\n- Returns 202 Accepted with task status","operationId":"generate_counter_offer_api_v1_counter_offer__clause_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"clause_id","in":"path","required":true,"schema":{"type":"string","title":"Clause Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["counter-offer"],"summary":"Get Counter Offer","description":"Polls for counter-offer result.\n\n- Returns 200 with data if ready\n- Returns 202 if still processing","operationId":"get_counter_offer_api_v1_counter_offer__clause_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"clause_id","in":"path","required":true,"schema":{"type":"string","title":"Clause Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/summary/{contract_id}":{"get":{"tags":["summary"],"summary":"Get Summary","description":"Returns the summary card and pros/cons for a contract.\n\n- Verifies JWT and ownership\n- Fetches summary from analysis_results table\n- Returns 404 if scan not complete or no summary exists","operationId":"get_summary_api_v1_summary__contract_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contract_id","in":"path","required":true,"schema":{"type":"string","title":"Contract Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/power/{contract_id}":{"get":{"tags":["power"],"summary":"Get Power Analysis","description":"Returns the power asymmetry analysis for a contract.\n\n- Verifies JWT and ownership\n- Fetches power analysis from analysis_results table\n- Returns 404 if scan not complete or no power analysis exists","operationId":"get_power_analysis_api_v1_power__contract_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contract_id","in":"path","required":true,"schema":{"type":"string","title":"Contract Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/precedent/{clause_id}":{"get":{"tags":["precedent"],"summary":"Get Precedent","description":"Returns the legal precedent match for a HIGH-risk clause.\n\n- Verifies JWT and ownership (clause must belong to user's contract)\n- Fetches precedent from precedent_matches table\n- Returns 404 if no precedent exists (e.g., for non-HIGH clauses)","operationId":"get_precedent_api_v1_precedent__clause_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"clause_id","in":"path","required":true,"schema":{"type":"string","title":"Clause Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/generate/{contractId}":{"post":{"tags":["report"],"summary":"Generate Report","description":"Verify ownership, create a report record, and queue the generation task.","operationId":"generate_report_api_v1_report_generate__contractId__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contractId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Contractid"}},{"name":"language","in":"query","required":false,"schema":{"type":"string","default":"en","title":"Language"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/{reportId}":{"get":{"tags":["report"],"summary":"Get Report","description":"Retrieve report details or download the PDF if it exists.","operationId":"get_report_api_v1_report__reportId__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"reportId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Reportid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/download/{reportId}":{"get":{"tags":["report"],"summary":"Download Report","description":"Internal download link for the authenticated user.","operationId":"download_report_api_v1_report_download__reportId__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"reportId","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Reportid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/share/{shareUuid}":{"get":{"tags":["report"],"summary":"Get Shared Report","description":"Public endpoint to access a report via share UUID.\nChecks for expiry.","operationId":"get_shared_report_api_v1_report_share__shareUuid__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"shareUuid","in":"path","required":true,"schema":{"type":"string","title":"Shareuuid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/{contract_id}":{"post":{"tags":["chat"],"summary":"Chat With Contract","description":"Stream chat response for a contract Q&A.\n\n- Verifies JWT and ownership\n- Checks that embeddings exist for the contract\n- Streams the answer via SSE","operationId":"chat_with_contract_api_v1_chat__contract_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contract_id","in":"path","required":true,"schema":{"type":"string","title":"Contract Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/translate/{contract_id}":{"post":{"tags":["translate"],"summary":"Translate Contract","description":"Queues translation of contract results to a different language.\n\n- Verifies JWT and ownership\n- Queues the translate_results Celery task\n- Returns 202 Accepted with task status","operationId":"translate_contract_api_v1_translate__contract_id__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"contract_id","in":"path","required":true,"schema":{"type":"string","title":"Contract Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ContractCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"contract_type":{"anyOf":[{"$ref":"#/components/schemas/ContractType"},{"type":"null"}]},"file_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Path"},"file_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Url"},"party_a":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Party A"},"party_b":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Party B"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","required":["name"],"title":"ContractCreate","description":"Request to create a new contract."},"ContractType":{"type":"string","enum":["Employment","NDA","Freelance/SOW","SaaS Subscription","Lease/Rental","Partnership","IP License","Loan","M&A","Other"],"title":"ContractType","description":"Contract types as defined in PRD Feature 2 (10 exact types)."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ScanResponse":{"properties":{"job_id":{"type":"string","format":"uuid","title":"Job Id"},"contract_id":{"type":"string","format":"uuid","title":"Contract Id"},"status":{"$ref":"#/components/schemas/ScanStatus"},"progress_pct":{"type":"number","title":"Progress Pct","default":0.0},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["job_id","contract_id","status"],"title":"ScanResponse","description":"Basic scan response."},"ScanStatus":{"type":"string","enum":["queued","processing","complete","failed","cancelled"],"title":"ScanStatus","description":"Status of a scan job."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}